requirements.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
future==0.18.2
idna==2.8
oauthlib==3.1.0
praw==6.5.1
prawcore==1.0.1
pycparser==2.19
PyMySQL==0.9.3
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.14.0
update-checker==0.16
urllib3==1.25.8
websocket-client==0.57.0
